+++ runtime/lib/utils.h
@@ -1,4 +1,5 @@
 #pragma once
+
 #include "core.h"
 
 #define CL_ERROR_BUF_SIZE 256
@@ -12,7 +13,7 @@ void cl_abort(char* message);
 
 void cl_abort_errno(char* message);
 
-void cl_check_func_args_count(const char* func_name, unsigned int args_count, unsigned int expected_count, enum CL_CountCheckingMode mode);
+void cl_check_func_args_count(const char* func_name, size_t args_count, size_t expected_count, enum CL_CountCheckingMode mode);
 
 void cl_check_func_arg_type(const char* func_name, enum CL_ObjectType type, enum CL_ObjectType expected_type);
